Fixed delete documents

+++ src/main/java/com/cencolshare/controller/DocumentController.java
@@ -16,9 +16,11 @@ import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.RedirectView;
 
 import com.cencolshare.model.Document;
+import com.cencolshare.model.Group;
 import com.cencolshare.model.Upload;
 import com.cencolshare.model.User;
 import com.cencolshare.service.DocumentService;
+import com.cencolshare.service.UploadService;
 import com.cencolshare.util.GroupUtil;
 
 @Controller
@@ -30,6 +32,9 @@ public class DocumentController extends BaseController {
 	
 	@Autowired
 	HttpServletRequest request;
+	
+	@Autowired
+	UploadService uploadService;
 	
 	@RequestMapping(value="/list", method=RequestMethod.GET)
 	public ModelAndView listDocument() {
@@ -57,6 +62,7 @@ public class DocumentController extends BaseController {
 		upload.setFileType(".pdf");
 		upload.setOriginalFileName("Java for beginner");
 		upload.setUploadDate(new Date());
+		Long uploadId = upload.getId();
 		
 		Document doc= new Document();
 		doc.setDocumentTitle(request.getParameter("documentTitle"));
@@ -66,10 +72,11 @@ public class DocumentController extends BaseController {
 		doc.setDateUploaded(new Date());
 		doc.setFileUrl("htttp://www.");
 		doc.setUpload(upload);
+		
 	
 		
 		if(!request.getParameter("documentId").equals("")) {
-			doc.setDocumentId(Integer.parseInt(request.getParameter("documentId")));
+			doc.setDocumentId(Long.parseLong(request.getParameter("documentId")));
 		}
 		
 		doc = documentService.saveDocument(doc);
@@ -80,8 +87,22 @@ public class DocumentController extends BaseController {
 	}
 	
 	@RequestMapping(value="/delete/{id}", method=RequestMethod.GET)
-	public ModelAndView deleteDocument(@PathVariable Integer id) {
-		documentService.deleteDocumentbyID(id);
+	public ModelAndView deleteDocument(@PathVariable Long id) {
+		
+		User user = getLoggedInUser();
+		final Document doc = documentService.getDocumentById(id);
+		Long uploadId = doc.getUpload().getId();
+		int docUserId = doc.getUser().getUserId();
+		int userID =user.getUserId();
+		
+		
+		if (userID == docUserId){
+		
+			documentService.deleteDocumentbyID(id);
+			uploadService.deleteUpload(uploadId);
+		}
+		//documentService.deleteDocumentbyID(id);
+		//uploadService.deleteUpload(uploadId);
 		return new ModelAndView(new RedirectView("/cencolshare/docs/list"));
 	}
